Move scrollIntoView into utils/scroll.ts
