+++ packages/utils/dom/index.ts
@@ -4,4 +4,3 @@ export * from './position'
 export * from './resize-event'
 export * from './scroll'
 export * from './style'
-export * from './scroll-into-view'
